Closes: #90
Made the destructor of the Cursor class more stable. It used to crash when an
error occured in the Cursor *constructor*.

Implemented a check that the parameter for Cursor() is actually an instance of
the Connection class (or a subclass thereof).


--HG--
extra : convert_revision : svn%3Af581f19b-c9e6-0310-bd56-e372eea7194d/pysqlite/trunk%40121