Closes: #99
In pysqlite 2.0.2, code like this would fail:

{{{
#!python
cur.fetchall()
cur.fetchall()
}

fetchmany() and even fetchone() could also fail if the statement was already
closed.

The change for prefetching one row in 2.0.2 was introducing a new problem: This
change also finalizes SQLite statements sooner - as soon as the last row was
prefetched, and not only when the next SQL is executed or the cursor object is
closed.

But there was also a check in cursor_iternext that checked if the SQLite
statement was available, if not it raised a ProgrammingError.

This check just had to be removed, and that's all that's changed pysqlite
2.0.3.


--HG--
extra : convert_revision : svn%3Af581f19b-c9e6-0310-bd56-e372eea7194d/pysqlite/trunk%40150